// src/app/model/user-info.ts
export class UserInfo {

    userId: number;
    enterpriseId: number;
    openid: string;
    nickname: string;
    age:number;
    sex: number;
    province: string;
    city: string;
    country: string;
    headimgurl: string;
    privilege: string;
    unionid: string;
    realName: string;
    phone: string;
    qq: string;
    speciality: string;
    major: string;
    schoolName: string;
    sort: number;
    status: number;
    subscribeStatus;
    isInfoComplete;
    createTime: Date;
    updateTime: Date;


}